requirements.txt
matplotlib ==3.3.3
streamlit ==0.72.0
Pillow
pandas
seaborn
pip install plotly ==4.0.0
requests